🐛 Fix paths to test files
pytest datafiles copies the given files into a temporary folder.
You have to add the datafiles argument to the test function and use it to define the final path in your test function
The lookup for the files to copy, starts at the root directory of your project.
So we had to replace `"testfiles/example_ahb.pdf",` with `"./unittests/testfiles/example_ahb.pdf",`
You can also leave out the `./` but it makes it more clear for the reader, where the path starts